requirements.txt
streamlit
supabase
python-dotenv
streamlit-extras
folium
streamlit_folium
pytest